Option to disable panel modifications
Added and option to prevent the extension from modifying the panel settings (transparency & background color). This allows the themes settings to remain while the blur background overlay still applied. This is useful for people who are using a theme with transparent/semi-transparent panels and only want to apply a blur to the background under the panel.