Fix the main menu favorites box from fading to blur
For some themes (i.e Mint-Y) there is a fade in time which caused the
favorites box to transition from a solid color into a shaded transparent
background. This change will over ride the theme's transition-duration
to be 0 so that the favorites box on the main-menu will instantly be
transparent.