Add notification support, disable panel effects on maximize option
Added the ability to blur the Notification Popup messages. This new Cinnamon component support is disabled by default, but when enabled, you will have the same control over the look of the Notification Popups that you already have over the other supported Cinnamon components.

Added an option "Restore theme settings when a window is maximized" under the Panels tab. This option (disabled by default) is intended to change the transparent panels back to their solid color background when a maximized window exists on the monitor where the panel exists. Thanks to Declan Snyder for contributing some code to enable this feature.

Added a welcome notification message that appear when the Blur Cinnamon extension is enabled. This notification has a button giving users easy access to the Blur Cinnamon configuration dialog window.

Added debugging messages to the BlurPopupMenus class (disabled by default)