Window effects, Focus effect, rounded corner support
1. Added the ability to apply (static) effects to almost any window. The window has to be (semi)transparent in order for the effects to be visible. There are a two ways to accomplish this: A) Many terminals programs support transparent backgrounds (i.e. Ptyxis, Tilix, Alcritty, Kitty, Konsole). When these terminals are configured to be transparent and the Blur Cinnamon Window effects are enabled for the terminal program, the effects will be visible. B) The Cinnamon extension called "Opacity Slider" allows you to make just about any window semitransparent and therefore allows the Blur Cinnamon window effect to be visible.

There are still some tracking issues that I have yet to find a way to fix. So when resizing some windows the blurred actor does not immediately follow the window resize, therefore you will see the blurred actor outside the window bounds for a fraction of a second. Hopefully, one day, I will find a better way to track resize events.

If you using the "Compiz Windows Effect" extension, the Blur Cinnamon window effects will "warp" in a way that is consistent with effects of the "Compiz Windows Effects" extension. So the blurred actor will not track the desktop background perfectly while the "wobble" effects are animating the window. It will match the background after the effects are complete. I intent to submit code to the "Compiz Windows Effect" extension which will add the ability to select applications where the effects will not apply, allowing you to disable the "wobble" effect for application windows where you have blur effects enabled.

As with all the current Blur Cinnamon effects, the window effects are static. This means that the effect shows a blurred copy of the desktop texture but any windows that are under a blurred window will not appear in the blur effect. Dynamic blurring is still something I intend to work on to see if I can find a way to implement for in Blur Cinnamon in the future.

2. I have added rounded corner effects to Blur Cinnamon. These effects are borrowed from Blur-my-shell, which borrowed from Rounded-window-corners which borrowed from Mutter. This allows Blur Cinnamon to allow rounded corners for Menus, Notifications, Tooltips and Windows. Blur Cinnamon will no longer disable rounded corners for Menus, Notifications and Tooltips when the theme uses rounded corners. It will (or should) adapt to rounded corner radius that the theme calls for. With application windows you need to specify which corners are rounded (top/bottom/both) and the radius needed.

3. Added a new "Focused window backlight effect" that will apply a "glow" backlight effect to the currently focused window. This effect uses a blurred background configured to spill over the window bounds creating a backling type effect that makes the focused window to appear to glow using colors from the desktop background image.

4. The Desktop effects now include the option to change the dimming (colorize) color.

5. Fixed (I believe) cases where the Tooltip blurred background would spill over the tooltip bounds